Move c8y mapper tests to c8y module and fix docs for tests.
