+++ crates/core/tedge_mapper/src/c8y/mod.rs
@@ -2,7 +2,6 @@ pub mod converter;
 pub mod error;
 mod fragments;
 pub mod mapper;
-mod operations;
 pub mod topic;
 
 #[cfg(test)]
