--- crates/core/tedge_mapper/src/c8y/operations.rs
@@ -1,219 +1,0 @@
-use crate::c8y::error::*;
-
-use c8y_smartrest::smartrest_deserializer::SmartRestLogRequest;
-use serde::{Deserialize, Serialize};
-use std::path::Path;
-use time::{format_description, OffsetDateTime};
-
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-/// used to retrieve the id of a log event
-pub struct SmartRestLogEvent {
-    pub id: String,
-}
-
-/// Returns a date time object from a file path or file-path-like string
-/// a typical file stem looks like this: "software-list-2021-10-27T10:29:58Z"
-///
-/// # Examples:
-/// ```
-/// let path_buf = PathBuf::fromStr("/path/to/file/with/date/in/path").unwrap();
-/// let path_bufdate_time = get_datetime_from_file_path(&path_buf).unwrap();
-/// ```
-fn get_datetime_from_file_path(log_path: &Path) -> Result<OffsetDateTime, CumulocityMapperError> {
-    if let Some(stem_string) = log_path.file_stem().and_then(|s| s.to_str()) {
-        // a typical file stem looks like this: software-list-2021-10-27T10:29:58Z.
-        // to extract the date, rsplit string on "-" and take (last) 3
-        let mut stem_string_vec = stem_string.rsplit('-').take(3).collect::<Vec<_>>();
-        // reverse back the order (because of rsplit)
-        stem_string_vec.reverse();
-        // join on '-' to get the date string
-        let date_string = stem_string_vec.join("-");
-        let dt = OffsetDateTime::parse(&date_string, &format_description::well_known::Rfc3339)?;
-
-        return Ok(dt);
-    }
-    match log_path.to_str() {
-        Some(path) => Err(CumulocityMapperError::InvalidDateInFileName(
-            path.to_string(),
-        )),
-        None => Err(CumulocityMapperError::InvalidUtf8Path),
-    }
-}
-
-/// Reads tedge logs according to `SmartRestLogRequest`.
-///
-/// If needed, logs are concatenated.
-///
-/// Logs are sorted alphanumerically from oldest to newest.
-///
-/// # Examples
-///
-/// ```
-/// let smartrest_obj = SmartRestLogRequest::from_smartrest(
-///     "522,DeviceSerial,syslog,2021-01-01T00:00:00+0200,2021-01-10T00:00:00+0200,,1000",
-/// )
-/// .unwrap();
-///
-/// let log = read_tedge_system_logs(&smartrest_obj, "/var/log/tedge").unwrap();
-/// ```
-fn read_tedge_logs(
-    smartrest_obj: &SmartRestLogRequest,
-    logs_dir: &str,
-) -> Result<String, CumulocityMapperError> {
-    let mut output = String::new();
-
-    // NOTE: As per documentation of std::fs::read_dir:
-    // "The order in which this iterator returns entries is platform and filesystem dependent."
-    // Therefore, files are sorted by date.
-    let mut read_vector: Vec<_> = std::fs::read_dir(logs_dir)?
-        .filter_map(|r| r.ok())
-        .filter_map(|dir_entry| {
-            let file_path = &dir_entry.path();
-            let datetime_object = get_datetime_from_file_path(file_path);
-            match datetime_object {
-                Ok(dt) => {
-                    if dt < smartrest_obj.date_from || dt > smartrest_obj.date_to {
-                        return None;
-                    }
-                    Some(dir_entry)
-                }
-                Err(_) => None,
-            }
-        })
-        .collect();
-
-    read_vector.sort_by_key(|dir| dir.path());
-
-    // loop sorted vector and push store log file to `output`
-    let mut line_counter: usize = 0;
-    for entry in read_vector {
-        let file_path = entry.path();
-        let file_content = std::fs::read_to_string(&file_path)?;
-        if file_content.is_empty() {
-            continue;
-        }
-
-        // adding file header only if line_counter permits more lines to be added
-        match &file_path.file_stem().and_then(|f| f.to_str()) {
-            Some(file_name) if line_counter < smartrest_obj.lines => {
-                output.push_str(&format!("filename: {}\n", file_name));
-            }
-            _ => {}
-        }
-
-        // split at new line delimiter ("\n")
-        let mut lines = file_content.lines();
-        while line_counter < smartrest_obj.lines {
-            if let Some(haystack) = lines.next() {
-                if let Some(needle) = &smartrest_obj.needle {
-                    if haystack.contains(needle) {
-                        output.push_str(&format!("{}\n", haystack));
-                        line_counter += 1;
-                    }
-                } else {
-                    output.push_str(&format!("{}\n", haystack));
-                    line_counter += 1;
-                }
-            } else {
-                // there are no lines.next()
-                break;
-            }
-        }
-    }
-    Ok(output)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Write;
-    use std::str::FromStr;
-    use std::{fs::File, path::PathBuf};
-    use test_case::test_case;
-
-    #[test_case("/path/to/software-list-2021-10-27T10:44:44Z.log")]
-    #[test_case("/path/to/tedge/agent/software-update-2021-10-25T07:45:41Z.log")]
-    #[test_case("/path/to/another-variant-2021-10-25T07:45:41Z.log")]
-    #[test_case("/yet-another-variant-2021-10-25T07:45:41Z.log")]
-    fn test_datetime_parsing_from_path(file_path: &str) {
-        // checking that `get_date_from_file_path` unwraps a `chrono::NaiveDateTime` object.
-        // this should return an Ok Result.
-        let path_buf = PathBuf::from_str(file_path).unwrap();
-        let path_buf_datetime = get_datetime_from_file_path(&path_buf);
-        assert!(path_buf_datetime.is_ok());
-    }
-
-    #[test_case("/path/to/software-list-2021-10-27-10:44:44Z.log")]
-    #[test_case("/path/to/tedge/agent/software-update-10-25-2021T07:45:41Z.log")]
-    #[test_case("/path/to/another-variant-07:45:41Z-2021-10-25T.log")]
-    #[test_case("/yet-another-variant-2021-10-25T07:45Z.log")]
-    fn test_datetime_parsing_from_path_fail(file_path: &str) {
-        // checking that `get_date_from_file_path` unwraps a `chrono::NaiveDateTime` object.
-        // this should return an err.
-        let path_buf = PathBuf::from_str(file_path).unwrap();
-        let path_buf_datetime = get_datetime_from_file_path(&path_buf);
-        assert!(path_buf_datetime.is_err());
-    }
-
-    fn parse_file_names_from_log_content(log_content: &str) -> [&str; 5] {
-        let mut files: Vec<&str> = vec![];
-        for line in log_content.lines() {
-            if line.contains("filename: ") {
-                let filename: &str = line.split("filename: ").last().unwrap();
-                files.push(filename);
-            }
-        }
-        match files.try_into() {
-            Ok(arr) => arr,
-            Err(_) => panic!("Could not convert to Array &str, size 5"),
-        }
-    }
-
-    #[test]
-    /// testing read_tedge_logs
-    ///
-    /// this test creates 5 fake log files in a temporary directory.
-    /// files are dated 2021-01-0XT01:00Z, where X = a different day.
-    ///
-    /// this tests will assert that files are read alphanumerically from oldest to newest
-    fn test_read_logs() {
-        // order in which files are created
-        const LOG_FILE_NAMES: [&str; 5] = [
-            "software-list-2021-01-03T01:00:00Z.log",
-            "software-list-2021-01-02T01:00:00Z.log",
-            "software-list-2021-01-01T01:00:00Z.log",
-            "software-update-2021-01-03T01:00:00Z.log",
-            "software-update-2021-01-02T01:00:00Z.log",
-        ];
-
-        // expected (sorted) output
-        const EXPECTED_OUTPUT: [&str; 5] = [
-            "software-list-2021-01-01T01:00:00Z",
-            "software-list-2021-01-02T01:00:00Z",
-            "software-list-2021-01-03T01:00:00Z",
-            "software-update-2021-01-02T01:00:00Z",
-            "software-update-2021-01-03T01:00:00Z",
-        ];
-
-        let smartrest_obj = SmartRestLogRequest::from_smartrest(
-            "522,DeviceSerial,syslog,2021-01-01T00:00:00+0200,2021-01-10T00:00:00+0200,,1000",
-        )
-        .unwrap();
-
-        let temp_dir = tempfile::tempdir().unwrap();
-        // creating the files
-        for (idx, file) in LOG_FILE_NAMES.iter().enumerate() {
-            let file_path = &temp_dir.path().join(file);
-            let mut file = File::create(file_path).unwrap();
-            writeln!(file, "file num {}", idx).unwrap();
-        }
-
-        // reading the logs and extracting the file names from the log output.
-        let output = read_tedge_logs(&smartrest_obj, temp_dir.path().to_str().unwrap()).unwrap();
-        let parsed_values = parse_file_names_from_log_content(&output);
-
-        // asserting the order = `EXPECTED_OUTPUT`
-        assert!(parsed_values.eq(&EXPECTED_OUTPUT));
-    }
-}
