+++ crates/core/tedge_mapper/src/c8y/tests.rs
@@ -1,8 +1,6 @@
-use crate::mapping::{
-    mapper::create_mapper, operations::Operations, size_threshold::SizeThreshold,
-};
-
+use crate::core::{mapper::create_mapper, size_threshold::SizeThreshold};
 use c8y_api::http_proxy::{C8YHttpProxy, FakeC8YHttpProxy, JwtAuthHttpProxy};
+use c8y_smartrest::operations::Operations;
 use mqtt_channel::{Connection, TopicFilter};
 use mqtt_tests::test_mqtt_server::MqttProcessHandler;
 use mqtt_tests::with_timeout::{Maybe, WithTimeout};
@@ -26,7 +24,7 @@ async fn mapper_publishes_a_software_list_request() {
         .await;
 
     // Start the SM Mapper
-    let sm_mapper = start_sm_mapper(broker.port).await;
+    let sm_mapper = start_c8y_mapper(broker.port).await;
 
     // Expect on `tedge/commands/req/software/list` a software list request.
     let msg = messages
@@ -48,7 +46,7 @@ async fn mapper_publishes_a_supported_operation_and_a_pending_operations_onto_c8
     let mut messages = broker.messages_published_on("c8y/s/us").await;
 
     // Start SM Mapper
-    let sm_mapper = start_sm_mapper(broker.port).await;
+    let sm_mapper = start_c8y_mapper(broker.port).await;
 
     let mut msg = messages
         .next()
@@ -86,7 +84,7 @@ async fn mapper_publishes_software_update_request() {
         .messages_published_on("tedge/commands/req/software/update")
         .await;
 
-    let sm_mapper = start_sm_mapper(broker.port).await;
+    let sm_mapper = start_c8y_mapper(broker.port).await;
 
     // Prepare and publish a software update smartrest request on `c8y/s/ds`.
     let smartrest = r#"528,external_id,nodered,1.0.0::debian,,install"#;
@@ -128,7 +126,7 @@ async fn mapper_publishes_software_update_status_onto_c8y_topic() {
     let mut messages = broker.messages_published_on("c8y/s/us").await;
 
     // Start SM Mapper
-    let sm_mapper = start_sm_mapper(broker.port).await;
+    let sm_mapper = start_c8y_mapper(broker.port).await;
     let _ = publish_a_fake_jwt_token(broker).await;
 
     let mut msg = messages
@@ -199,7 +197,7 @@ async fn mapper_publishes_software_update_failed_status_onto_c8y_topic() {
     let mut messages = broker.messages_published_on("c8y/s/us").await;
 
     // Start SM Mapper
-    let sm_mapper = start_sm_mapper(broker.port).await;
+    let sm_mapper = start_c8y_mapper(broker.port).await;
     let _ = publish_a_fake_jwt_token(broker).await;
 
     let mut msg = messages
@@ -280,7 +278,7 @@ async fn mapper_fails_during_sw_update_recovers_and_process_response() -> Result
     let mut responses = broker.messages_published_on("c8y/s/us").await;
 
     // Start SM Mapper
-    let sm_mapper = start_sm_mapper(broker.port).await?;
+    let sm_mapper = start_c8y_mapper(broker.port).await?;
 
     let mut msg = responses
         .next()
@@ -351,7 +349,7 @@ async fn mapper_fails_during_sw_update_recovers_and_process_response() -> Result
         .unwrap();
 
     // Restart SM Mapper
-    let sm_mapper = start_sm_mapper(broker.port).await?;
+    let sm_mapper = start_c8y_mapper(broker.port).await?;
 
     // Validate that the mapper process the response and forward it on 'c8y/s/us'
     // Expect init messages followed by a 503 (success)
@@ -384,7 +382,7 @@ async fn mapper_publishes_software_update_request_with_wrong_action() {
     // Create a subscriber to receive messages on `c8y/s/us` topic.
     let mut messages = broker.messages_published_on("c8y/s/us").await;
 
-    let _sm_mapper = start_sm_mapper(broker.port).await;
+    let _sm_mapper = start_c8y_mapper(broker.port).await;
 
     let mut msg = messages
         .next()
@@ -449,40 +447,167 @@ async fn get_jwt_token_full_run() {
     assert_eq!(jwt_token.unwrap().token(), "1111");
 }
 
-fn remove_whitespace(s: &str) -> String {
-    let mut s = String::from(s);
-    s.retain(|c| !c.is_whitespace());
-    s
+const ALARM_SYNC_TIMEOUT_MS: Duration = Duration::from_millis(5000);
+
+#[tokio::test]
+#[serial]
+async fn c8y_mapper_alarm_mapping_to_smartrest() {
+    let broker = mqtt_tests::test_mqtt_broker();
+
+    let mut messages = broker.messages_published_on("c8y/s/us").await;
+
+    // Start the C8Y Mapper
+    let c8y_mapper = start_c8y_mapper(broker.port).await.unwrap();
+
+    let _ = broker
+        .publish_with_opts(
+            "tedge/alarms/major/temperature_alarm",
+            r#"{ "message": "Temperature high" }"#,
+            mqtt_channel::QoS::AtLeastOnce,
+            true,
+        )
+        .await
+        .unwrap();
+
+    let mut msg = messages
+        .next()
+        .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+        .await
+        .expect_or("No message received before timeout");
+    dbg!(&msg);
+
+    // The first message could be SmartREST 114 for supported operations
+    while !msg.contains("302") {
+        // Fetch the next message which should be the alarm
+        msg = messages
+            .next()
+            .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+            .await
+            .expect_or("No message received before timeout");
+    }
+
+    // Expect converted temperature alarm message
+    dbg!(&msg);
+    assert!(msg.contains("302,temperature_alarm"));
+
+    //Clear the previously published alarm
+    let _ = broker
+        .publish_with_opts(
+            "tedge/alarms/major/temperature_alarm",
+            "",
+            mqtt_channel::QoS::AtLeastOnce,
+            true,
+        )
+        .await
+        .unwrap();
+
+    c8y_mapper.abort();
 }
 
-// async fn start_sm_mapper(mqtt_port: u16) -> Result<JoinHandle<()>, anyhow::Error> {
-//     let operations = Operations::new();
-//     let mqtt_topic = CumulocityMapper::subscriptions(&operations)?;
-//     let mqtt_config = mqtt_channel::Config::default()
-//         .with_port(mqtt_port)
-//         .with_session_name("SM-C8Y-Mapper-Test")
-//         .with_subscriptions(mqtt_topic);
+#[tokio::test]
+#[serial]
+async fn c8y_mapper_syncs_pending_alarms_on_startup() {
+    let broker = mqtt_tests::test_mqtt_broker();
 
-//     let mqtt_client = Connection::new(&mqtt_config).await?;
-//     let http_proxy = FakeC8YHttpProxy {};
-//     let device_name = "test-device".into();
-//     let device_type = "test-device-type".into();
-//     let size_threshold = SizeThreshold(16 * 1024);
-//     let mut mapper = CumulocityConverter::new(
-//         size_threshold,
-//         device_name,
-//         device_type,
-//         operations,
-//         http_proxy,
-//     );
+    let mut messages = broker.messages_published_on("c8y/s/us").await;
 
-//     let mapper_task = tokio::spawn(async move {
-//         let _ = mapper.run().await;
-//     });
-//     Ok(mapper_task)
-// }
+    // Start the C8Y Mapper
+    let c8y_mapper = start_c8y_mapper(broker.port).await.unwrap();
 
-async fn start_sm_mapper(mqtt_port: u16) -> Result<JoinHandle<()>, anyhow::Error> {
+    let _ = broker
+        .publish_with_opts(
+            "tedge/alarms/critical/temperature_alarm",
+            r#"{ "message": "Temperature very high" }"#,
+            mqtt_channel::QoS::AtLeastOnce,
+            true,
+        )
+        .await
+        .unwrap();
+
+    let mut msg = messages
+        .next()
+        .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+        .await
+        .expect_or("No message received before timeout");
+    dbg!(&msg);
+
+    // The first message could be SmartREST 114 for supported operations
+    while !msg.contains("301") {
+        // Fetch the next message which should be the alarm
+        msg = messages
+            .next()
+            .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+            .await
+            .expect_or("No message received before timeout");
+        dbg!(&msg);
+    }
+
+    // Expect converted temperature alarm message
+    assert!(&msg.contains("301,temperature_alarm"));
+
+    c8y_mapper.abort();
+
+    //Publish a new alarm while the mapper is down
+    let _ = broker
+        .publish_with_opts(
+            "tedge/alarms/critical/pressure_alarm",
+            r#"{ "message": "Pressure very high" }"#,
+            mqtt_channel::QoS::AtLeastOnce,
+            true,
+        )
+        .await
+        .unwrap();
+
+    // Ignored until the rumqttd broker bug that doesn't handle empty retained messages
+    //Clear the existing alarm while the mapper is down
+    // let _ = broker
+    //     .publish_with_opts(
+    //         "tedge/alarms/critical/temperature_alarm",
+    //         "",
+    //         mqtt_channel::QoS::AtLeastOnce,
+    //         true,
+    //     )
+    //     .await
+    //     .unwrap();
+
+    // Restart the C8Y Mapper
+    let _ = start_c8y_mapper(broker.port).await.unwrap();
+
+    let mut msg = messages
+        .next()
+        .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+        .await
+        .expect_or("No message received before timeout");
+    dbg!(&msg);
+
+    // The first message could be SmartREST 114 for supported operations
+    while !msg.contains("301") {
+        // Fetch the next message which should be the alarm
+        msg = messages
+            .next()
+            .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+            .await
+            .expect_or("No message received before timeout");
+        dbg!(&msg);
+    }
+
+    // Ignored until the rumqttd broker bug that doesn't handle empty retained messages
+    // Expect the previously missed clear temperature alarm message
+    // let msg = messages
+    //     .next()
+    //     .with_timeout(ALARM_SYNC_TIMEOUT_MS)
+    //     .await
+    //     .expect_or("No message received after a second.");
+    // dbg!(&msg);
+    // assert!(&msg.contains("306,temperature_alarm"));
+
+    // Expect the new pressure alarm message
+    assert!(&msg.contains("301,pressure_alarm"));
+
+    c8y_mapper.abort();
+}
+
+async fn start_c8y_mapper(mqtt_port: u16) -> Result<JoinHandle<()>, anyhow::Error> {
     let device_name = "test-device".into();
     let device_type = "test-device-type".into();
     let size_threshold = SizeThreshold(16 * 1024);
@@ -503,6 +628,12 @@ async fn start_sm_mapper(mqtt_port: u16) -> Result<JoinHandle<()>, anyhow::Error
         let _ = mapper.run().await;
     });
     Ok(mapper_task)
+}
+
+fn remove_whitespace(s: &str) -> String {
+    let mut s = String::from(s);
+    s.retain(|c| !c.is_whitespace());
+    s
 }
 
 async fn publish_a_fake_jwt_token(broker: &MqttProcessHandler) {
