+++ crates/core/tedge_mapper/src/core/component.rs
