+++ crates/core/tedge_mapper/src/core/error.rs
@@ -1,7 +1,7 @@
-use crate::mapping::size_threshold::SizeThresholdExceeded;
+use crate::core::size_threshold::SizeThresholdExceeded;
 
+use c8y_smartrest::error::OperationsError;
 use mqtt_channel::MqttError;
-use std::path::PathBuf;
 use tedge_config::TEdgeConfigError;
 use thin_edge_json::serialize::ThinEdgeJsonSerializationError;
 
@@ -76,16 +76,4 @@ pub enum ConversionError {
 
     #[error(transparent)]
     FromUtf8Error(#[from] std::string::FromUtf8Error),
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum OperationsError {
-    #[error(transparent)]
-    FromIo(#[from] std::io::Error),
-
-    #[error("Cannot extract the operation name from the path: {0}")]
-    InvalidOperationName(PathBuf),
-
-    #[error("Error while parsing operation file: '{0}': {1}.")]
-    TomlError(PathBuf, #[source] toml::de::Error),
 }
