+++ crates/core/tedge_mapper/src/core/mod.rs
@@ -2,5 +2,4 @@ pub mod component;
 pub mod converter;
 pub mod error;
 pub mod mapper;
-pub mod operations;
 pub mod size_threshold;
